backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==3.2.16
django-extra-fields==3.0.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
gunicorn==20.1.0
isort==5.10.1
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.8.6
PyJWT==2.6.0
python-dotenv==0.20.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.13
sqlparse==0.4.3